Migrate from deprecated SimpleClient to KafkaClient
kafka-python’s SimpleClient has been deprecated for a while in favor of the new KafkaClient. Switching to the new client inherits much better error handling and retries of transient errors from upstream kaka-python.

It also allows for  executing this check faster by making calls to each broker in the cluster in parallel, but for simplicity I implemented the calls in a blocking manner.

Lastly, switching to the new Client exposed the error codes for each partition so I added additional error handling to handle when the partition leader is switching to a different broker.